chore(3040): review fixies, projection for user collectio query
